Update: 3. Bug fixed: Modify some parts in run.sh

+++ Lab3/run.sh
@@ -6,14 +6,6 @@ src_name="Test"
 target_name="Test_IR"
 file_num=$(find ./$src_name/ -name "*.txt" | wc -l)
 
-# Find the target storage folder. If it doesn't exit, create one.
-if [ $(find ./ -name ${target_name} | wc -l) != 0 ];then
-    rm -rf ${target_name}
-    mkdir ${target_name}
-else
-    mkdir ${target_name}
-fi
-
 # Default source file path will be told to users.
 echo "Default source file path: ${src_name}"
 echo "If you want to chage the source file path, please edit the file called \"run.sh\""
@@ -41,6 +33,7 @@ else
     cd Code
     make clean
     make parser
+
     # Change the location of the excutable file.
     if [ $(find ./ -name "parser" | wc -l) != 0 ];then
         cp parser ./../parser
@@ -50,37 +43,51 @@ else
 
     # Analyze the parameters and output the results.
     if [ $paras_num == 0 ];then
-    echo "Default Mode: generate the intermediate representation code"
-    echo "Default Results Storage Path: ${target_name}"
-    for ((i=1;i<=$file_num;i++))
-    do
-        ./parser -ir ./$src_name/test${i}.txt 1>>${target_name}/test${i}_ir.ir 2>>${target_name}/test${i}_ir.ir
-    done
+        echo "Default Mode: generate the intermediate representation code"
+        echo "Default Results Storage Path: ${target_name}"
+        
+        # Find the target storage folder. If it doesn't exit, create one.
+        if [ $(find ./ -name ${target_name} | wc -l) != 0 ];then
+            rm -rf ${target_name}
+            mkdir ${target_name}
+        else
+            mkdir ${target_name}
+        fi
+
+        # Start the program execution
+        for ((i=1;i<=$file_num;i++))
+        do
+            ./parser -ir ./$src_name/test${i}.txt 1>>${target_name}/test${i}_ir.ir 2>>${target_name}/test${i}_ir.ir
+        done
     else
-    rm -rf ${target_name}
-    args=" "
-    flag=0
-    echo "User-Chosen Mode: " $@
-    for arg in "$@"
-    do
-        if [ ${arg} == "-p" ];then
-            flag=1
-        fi
-        if [ ${arg:0:1} != "-" ];then
-            target_name=${arg}
-        else
-            args=${args}" "${arg}
-        fi
-    done
-    echo "User-Chosen Storage Path: ${target_name}"
-    mkdir ${target_name}
-    for ((i=1;i<=$file_num;i++))
-    do
-        if [ $flag == 0 ];then
-            ./parser ${args} ./$src_name/test${i}.txt 1>>${target_name}/test${i}_ir.ir 2>>${target_name}/test${i}_ir.ir
-        else
-            ./parser ${args} ./$src_name/test${i}.txt 1>>${target_name}/test${i}_st_ir.txt 2>>${target_name}/test${i}_st_ir.txt
-        fi
-    done
+        rm -rf ${target_name}
+        args=" "
+        flag=0
+        echo "User-Chosen Mode: " $@
+
+        # Determine the result path and other modes that are chosen
+        for arg in "$@"
+        do
+            if [ ${arg} == "-p" ];then
+                flag=1
+            fi
+            if [ ${arg:0:1} != "-" ];then
+                target_name=${arg}
+            else
+                args=${args}" "${arg}
+            fi
+        done
+
+        # Start the program execution
+        echo "User-Chosen Storage Path: ${target_name}"
+        mkdir ${target_name}
+        for ((i=1;i<=$file_num;i++))
+        do
+            if [ $flag == 0 ];then
+                ./parser ${args} ./$src_name/test${i}.txt 1>>${target_name}/test${i}_ir.ir 2>>${target_name}/test${i}_ir.ir
+            else
+                ./parser ${args} ./$src_name/test${i}.txt 1>>${target_name}/test${i}_st_ir.txt 2>>${target_name}/test${i}_st_ir.txt
+            fi
+        done
     fi
 fi
